Move the logic to operators.python
