+++ airflow/models/taskinstance.py
@@ -15,6 +15,7 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+import contextlib
 import getpass
 import hashlib
 import logging
@@ -51,7 +52,6 @@ from airflow.models.xcom import XCOM_RETURN_KEY, XCom
 from airflow.sentry import Sentry
 from airflow.settings import STORE_SERIALIZED_DAGS
 from airflow.stats import Stats
-from airflow.task.context.current import set_current_context
 from airflow.ti_deps.dep_context import DepContext
 from airflow.ti_deps.dependencies_deps import REQUEUEABLE_DEPS, RUNNING_DEPS
 from airflow.utils import timezone
@@ -66,6 +66,28 @@ from airflow.utils.state import State
 from airflow.utils.timeout import timeout
 
 TR = TaskReschedule
+
+_CURRENT_CONTEXT = []
+log = logging.getLogger(__name__)
+
+
+@contextlib.contextmanager
+def set_current_context(context: Dict[str, Any]):
+    """
+    Sets the current execution context to the provided context object.
+    This method should be called once per Task execution, before calling operator.execute.
+    """
+    _CURRENT_CONTEXT.append(context)
+    try:
+        yield context
+    finally:
+        expected_state = _CURRENT_CONTEXT.pop()
+        if expected_state != context:
+            log.warning(
+                "Current context is not equal to the state at context stack. Expected=%s, got=%s",
+                context,
+                expected_state,
+            )
 
 
 def clear_task_instances(tis,
