+++ docs/concepts.rst
@@ -278,8 +278,7 @@ using ``@task`` decorator.
 
 .. code-block:: python
 
-    from airflow.decorators import task
-    from airflow.task.context import get_current_context
+    from airflow.operators.python import task, get_current_context
 
     @task
     def my_task():
