--- tests/task/context/test_current_context.py
@@ -1,109 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-from datetime import datetime, timedelta
-
-import pytest
-
-from airflow import DAG
-from airflow.exceptions import AirflowException
-from airflow.models.baseoperator import BaseOperator
-from airflow.operators.python import PythonOperator
-from airflow.task.context.current import get_current_context, set_current_context
-from airflow.utils.dates import days_ago
-from tests.test_utils.db import clear_db_runs
-
-DEFAULT_ARGS = {
-    "owner": "test",
-    "depends_on_past": True,
-    "start_date": days_ago(1),
-    "end_date": datetime.today(),
-    "schedule_interval": "@once",
-    "retries": 1,
-    "retry_delay": timedelta(minutes=1),
-}
-
-
-class TestCurrentContext:
-    def test_current_context_no_context_raise(self):
-        with pytest.raises(AirflowException):
-            get_current_context()
-
-    def test_current_context_roundtrip(self):
-        example_context = {"Hello": "World"}
-
-        with set_current_context(example_context):
-            assert get_current_context() == example_context
-
-    def test_context_removed_after_exit(self):
-        example_context = {"Hello": "World"}
-
-        with set_current_context(example_context):
-            pass
-        with pytest.raises(AirflowException, ):
-            get_current_context()
-
-    def test_nested_context(self):
-        """
-        Nested execution context should be supported in case the user uses multiple context managers.
-        Each time the execute method of an operator is called, we set a new 'current' context.
-        This test verifies that no matter how many contexts are entered - order is preserved
-        """
-        max_stack_depth = 15
-        ctx_list = []
-        for i in range(max_stack_depth):
-            # Create all contexts in ascending order
-            new_context = {"ContextId": i}
-            # Like 15 nested with statements
-            ctx_obj = set_current_context(new_context)
-            ctx_obj.__enter__()  # pylint: disable=E1101
-            ctx_list.append(ctx_obj)
-        for i in reversed(range(max_stack_depth)):
-            # Iterate over contexts in reverse order - stack is LIFO
-            ctx = get_current_context()
-            assert ctx["ContextId"] == i
-            # End of with statement
-            ctx_list[i].__exit__(None, None, None)
-
-
-class MyContextAssertOperator(BaseOperator):
-    def execute(self, context):
-        assert context == get_current_context()
-
-
-def get_all_the_context(**context):
-    current_context = get_current_context()
-    assert context == current_context
-
-
-@pytest.fixture()
-def clear_db():
-    clear_db_runs()
-    yield
-    clear_db_runs()
-
-
-@pytest.mark.usefixtures("clear_db")
-class TestCurrentContextRuntime:
-    def test_context_in_task(self):
-        with DAG(dag_id="assert_context_dag", default_args=DEFAULT_ARGS):
-            op = MyContextAssertOperator(task_id="assert_context")
-            op.run(ignore_first_depends_on_past=True, ignore_ti_state=True)
-
-    def test_get_context_in_old_style_context_task(self):
-        with DAG(dag_id="edge_case_context_dag", default_args=DEFAULT_ARGS):
-            op = PythonOperator(python_callable=get_all_the_context, task_id="get_all_the_context")
-            op.run(ignore_first_depends_on_past=True, ignore_ti_state=True)
